refactor: table <-> order 의존관계 inteface를 통해서 제거
