+++ kitchenpos-order/src/main/java/kitchenpos/table/dto/TableRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.dto;
+package kitchenpos.table.dto;
 
 public class TableRequest {
     private int numberOfGuests;
