+++ kitchenpos-order/src/main/java/kitchenpos/table/dto/TableResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.dto;
+package kitchenpos.table.dto;
 
 import kitchenpos.table.domain.OrderTable;
