+++ kitchenpos-order/src/main/java/kitchenpos/table/port/OrderTableValidatorPort.java
@@ -1,0 +1,8 @@
+package kitchenpos.table.port;
+
+import kitchenpos.table.domain.OrderTable;
+
+public interface OrderTableValidatorPort {
+    void validChangeEmpty(OrderTable orderTable);
+    void validChangeNumberOfGuest(int numberOfGuests);
+}
