+++ kitchenpos-order/src/test/java/kitchenpos/acceptence/TableRestControllerTest.java
@@ -6,8 +6,8 @@ import io.restassured.response.Response;
 import kitchenpos.AcceptanceSupport;
 import kitchenpos.order.dto.ChaneNumberOfGuestRequest;
 import kitchenpos.order.dto.ChangeEmptyRequest;
-import kitchenpos.order.dto.TableRequest;
-import kitchenpos.order.dto.TableResponse;
+import kitchenpos.table.dto.TableRequest;
+import kitchenpos.table.dto.TableResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
