+++ kitchenpos-order/src/test/java/kitchenpos/unit/TableServiceTest.java
@@ -5,8 +5,8 @@ import kitchenpos.table.domain.OrderTable;
 import kitchenpos.order.validator.OrderTableValidator;
 import kitchenpos.order.dto.ChaneNumberOfGuestRequest;
 import kitchenpos.order.dto.ChangeEmptyRequest;
-import kitchenpos.order.dto.TableRequest;
-import kitchenpos.order.dto.TableResponse;
+import kitchenpos.table.dto.TableRequest;
+import kitchenpos.table.dto.TableResponse;
 import kitchenpos.table.port.OrderTablePort;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
